refs #75 Project#FetchGithubとhook用のメソッド群を共通化する
